requirements.txt
# ─── AudiobookMaker — Python Dependencies ────────────────────────────────────
# Install with:  pip install -r requirements.txt
# Python 3.10+ required

# ── Core book/text extraction ─────────────────────────────────────────────────
EbookLib            # EPUB read/write (.epub, .mobi)
beautifulsoup4      # HTML parsing
lxml                # Fast HTML/XML parser (BS4 backend)
docling             # AI-powered document extraction (PDF, EPUB, DOCX)
pymupdf             # PDF page extraction + TOC (fitz)
python-docx         # DOCX extraction
odfpy               # ODT extraction
docx2txt            # Lightweight DOCX text extraction fallback

# ── OCR ───────────────────────────────────────────────────────────────────────
easyocr             # Image text extraction inside EPUBs
pillow              # Image decoding for EasyOCR
numpy               # Array ops for EasyOCR / audio processing

# ── AI / ML classifier (optional) ────────────────────────────────────────────
xgboost             # Chapter classifier (MLClassifier in extractor_engine.py)

# ── TTS (Qwen3-TTS) ──────────────────────────────────────────────────────────
transformers        # HuggingFace model loading
torch               # PyTorch backend
safetensors         # Fast model weight loading
tokenizers          # Fast tokenizer
accelerate          # Multi-GPU / CPU offload
huggingface-hub     # Model download

# ── BookNLP (story / character analysis) ─────────────────────────────────────
booknlp

# ── Audio processing ──────────────────────────────────────────────────────────
soundfile           # Read/write WAV/FLAC
librosa             # Audio features (RMS, resample, etc.)
noisereduce         # Noise reduction for voice preprocessing
scipy               # Butterworth high-pass filter
mutagen             # Audio metadata tagging (ID3, MP4)
pyloudnorm          # LUFS-based loudness normalization

# ── UI ────────────────────────────────────────────────────────────────────────
gradio              # Web UI

# ── Utilities ─────────────────────────────────────────────────────────────────
requests
tqdm
pydantic
rich                # Pretty console output
